# api/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "semrel registry API",
    "version": "1.0.0",
    "description": "The semrel plugin registry. The public catalogue is readable without authentication; publishing and review require a GitHub session.\n\nBrowsers authenticate with an HttpOnly `semrel_session` cookie set by the OAuth callback. API clients (the semrel CLI, CI jobs) send `Authorization: Bearer <token>`.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "https://registry.semrel.io",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/plugins.json": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "The full plugin catalogue",
        "description": "The index the semrel CLI consumes. Carries a strong ETag; send If-None-Match to get a 304 instead of the whole document.",
        "security": [],
        "parameters": [
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "A previously returned ETag."
          }
        ],
        "responses": {
          "200": {
            "description": "The catalogue.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Strong validator over the response body."
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalogue"
                }
              }
            }
          },
          "304": {
            "description": "The client's cached copy is current."
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/api/v1/plugins": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "List plugins",
        "description": "Anonymous callers see active plugins. An authenticated non-admin additionally sees their own plugins in every state; an admin may filter by status.",
        "security": [
          {},
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Category"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Full-text and partial-identifier search. Results are ranked by relevance unless an explicit sort is given."
          },
          {
            "name": "compatibleWith",
            "in": "query",
            "schema": {
              "type": "string",
              "examples": [
                "0.27.1"
              ]
            },
            "description": "A concrete semrel core version. Returns only plugins whose latest installable release declares a compatibility range admitting it. Plugins that declare no range are included: the metadata is optional and most of the catalogue predates it."
          },
          {
            "name": "author",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Status"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "category",
                "created_at",
                "updated_at",
                "downloads",
                "views"
              ]
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of plugins.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      },
      "post": {
        "tags": [
          "publishing"
        ],
        "summary": "Create a plugin",
        "description": "Non-admin callers have the author forced to their own GitHub login.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "semrel_session",
        "description": "HttpOnly, SameSite=Lax session cookie set by the OAuth callback. Browser clients use this; mutating requests are additionally checked against the Origin header."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "For non-browser clients. The static ADMIN_TOKEN is accepted outside production only."
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Hub-Signature-256",
        "description": "HMAC-SHA256 over the raw request body, keyed with WEBHOOK_SECRET, formatted as sha256=<hex>."
      }
    },
...
    "responses": {
      "ValidationError": {
        "description": "The request was rejected by validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required, or the session has expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated, but not permitted. Publishers may only touch their own plugins.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "Too many requests. X-RateLimit-Limit and X-RateLimit-Remaining describe the budget.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          },
          "X-RateLimit-Limit": {
            "schema": {
              "type": "string"
            }
          },
          "X-RateLimit-Remaining": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Category": {
        "type": "string",
        "enum": [
          "analyzer",
          "condition",
          "generator",
          "hook",
          "provider",
          "updater",
          "packager",
          "publisher"
        ]
      },
      "Status": {
        "type": "string",
        "enum": [
          "active",
          "pending",
          "rejected"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "examples": [
                  "VALIDATION_ERROR",
                  "REAUTH_REQUIRED",
                  "ORIGIN_NOT_ALLOWED"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "Plugin": {
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "namespace": {
            "type": "string",
            "examples": [
              "@semrel"
            ]
          },
          "name": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "repository": {
            "type": "string",
            "format": "uri"
          },
          "license": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "latestVersion": {
            "type": "string",
            "description": "The newest stable, non-yanked version."
          },
          "views": {
            "type": "integer",
            "format": "int64"
          },
          "downloads": {
            "type": "integer",
            "format": "int64"
          },
          "rejectionReason": {
            "type": "string",
            "description": "Why a submission was rejected. Shown to the author."
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedBy": {
            "type": "string"
          },
          "validatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "securityAdvisories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecurityAdvisory"
            }
          },
          "advisoriesCheckedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When security advisories were last imported. Absent means never attempted."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "latestSemrelCore": {
            "type": "string",
            "description": "The semrel core compatibility range of latestVersion."
          }
        }
      },
      "PluginInput": {
        "type": "object",
        "required": [
          "name",
          "category"
        ],
        "properties": {
          "namespace": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "maxLength": 4000
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "repository": {
            "type": "string",
            "format": "uri"
          },
          "license": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer"
          }
        }
      },
      "PluginList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plugin"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "PluginEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Plugin"
          }
        }
      },
...
      "Catalogue": {
        "type": "object",
        "description": "The registry index. Its authoritative schema is schemas/plugin-metadata.json.",
        "properties": {
          "schemaVersion": {
            "type": "integer"
          },
          "plugins": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "SecurityAdvisory": {
        "type": "object",
        "description": "A known vulnerability affecting a plugin, imported from the GitHub security advisories published on the plugin's own repository.",
        "properties": {
          "ghsaId": {
            "type": "string",
            "examples": [
              "GHSA-xxxx-xxxx-xxxx"
            ]
          },
          "cveId": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "vulnerableRange": {
            "type": "string",
            "description": "Affected versions, in this registry's own range syntax (see the version schema's compatibility.semrelCore)."
          },
          "patchedVersion": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "withdrawnAt": {
            "type": "string",
            "format": "date-time",
            "description": "Set when GitHub itself retracted the advisory, e.g. a false positive."
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerAuth": []
    }
  ]
}